Create SemanticChatMemorySkill
